+++ NEWS.md
@@ -7,6 +7,9 @@
 * create_datastore_script modified to allow upload of large (>32mb) repos
 
 # QCkit v1.1.0
+## 2025-07-25
+* add function `get_user_orcid`, which will look up a users orcid using active directory. 
+
 ## 2025-05-16
 * create_datastore_script now invisibly returns the DataStore reference URL